#1 - update dependencies and create qs-api with schema,  dto generation, and jakarta interfaces for rest.
