+++ qs-api/qs-api-dto/src/main/resources/xsd/hu/icellmobilsoft/quarkus/sampler/dto/bindings.xjb
