+++ qs-api/qs-api-jakarta/src/main/java/hu/icellmobilsoft/quarkus/sampler/api/jakarta/path/ServicePath.java
@@ -1,0 +1,48 @@
+/*-
+ * #%L
+ * Quarkus-sampler
+ * %%
+ * Copyright (C) 2024 i-Cell Mobilsoft Zrt.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package hu.icellmobilsoft.quarkus.sampler.api.jakarta.path;
+
+import hu.icellmobilsoft.coffee.dto.url.BaseServicePath;
+
+/**
+ * Project specific rest services path container
+ *
+ * @author speter555
+ * @since 0.1.0
+ */
+public class ServicePath extends BaseServicePath {
+
+    /**
+     * {@value} prefix
+     */
+    public static final String TEST = "/test";
+
+    /**
+     * VersionInfo postfix path
+     */
+    public static final String VERSION_INFO = "/versionInfo";
+
+    /**
+     * Default constructor
+     */
+    public ServicePath() {
+        // Default constructor for java 21
+    }
+}
