+++ qs-api/qs-api-schema/src/main/java/hu/icellmobilsoft/quarkus/sampler/api/schema/XsdConstants.java
@@ -1,0 +1,36 @@
+/*-
+ * #%L
+ * Quarkus-sampler
+ * %%
+ * Copyright (C) 2024 i-Cell Mobilsoft Zrt.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package hu.icellmobilsoft.quarkus.sampler.api.schema;
+
+/**
+ * Constants for xsd configuration
+ *
+ * @author speter555
+ *
+ * @since 0.1.0
+ */
+public interface XsdConstants {
+
+    /**
+     * Path of the super XSD
+     */
+    String SUPER_XSD_PATH = "xsd/hu/icellmobilsoft/quarkus/sampler/dto/super.xsd";
+
+}
