+++ qs-api/qs-api-schema/src/main/resources/xsd/hu/icellmobilsoft/quarkus/sampler/dto/test/post.xsd
