+++ qs-common/pom.xml
@@ -13,8 +13,6 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>qs-common-api</module>
-        <module>qs-common-dto</module>
         <module>qs-common-rest</module>
         <module>qs-common-health</module>
         <module>qs-common-core</module>
