--- qs-common/qs-common-api/src/main/java/hu/icellmobilsoft/quarkus/sampler/api/system/ISystemRest.java
@@ -1,53 +1,0 @@
-/*-
- * #%L
- * Quarkus-sampler
- * %%
- * Copyright (C) 2024 i-Cell Mobilsoft Zrt.
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package hu.icellmobilsoft.quarkus.sampler.api.system;
-
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-
-import org.eclipse.microprofile.openapi.annotations.Operation;
-
-import hu.icellmobilsoft.coffee.dto.exception.BaseException;
-import hu.icellmobilsoft.quarkus.sampler.dto.path.ServicePath;
-
-/**
- * REST endpoint for system service functions.
- *
- * @author speter555
- * @since 0.1.0
- */
-@Path("")
-public interface ISystemRest {
-
-    /**
-     * Version info
-     * 
-     * @return version infos in plain text
-     * @throws BaseException
-     *             if any error occurs
-     */
-    @Operation(hidden = true)
-    @GET
-    @Path(ServicePath.VERSION_INFO)
-    @Produces(MediaType.TEXT_PLAIN)
-    String versionInfo() throws BaseException;
-}
