--- qs-common/qs-common-dto/pom.xml
@@ -1,131 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>hu.icellmobilsoft.quarkus.sampler</groupId>
-        <artifactId>qs-common</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>qs-common-dto</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>hu.icellmobilsoft.coffee</groupId>
-            <artifactId>coffee-dto-gen</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.xml.bind</groupId>
-            <artifactId>jakarta.xml.bind-api</artifactId>
-            <version>4.0.0</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.activation</groupId>
-            <artifactId>jakarta.activation-api</artifactId>
-            <version>2.1.0</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Data bind -->
-
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jackson</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jaxb</artifactId>
-        </dependency>
-    </dependencies>
-
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>com.evolvedbinary.maven.jvnet</groupId>
-                <artifactId>jaxb30-maven-plugin</artifactId>
-                <version>0.15.0</version>
-                <executions>
-                    <execution>
-                        <id>icell-super</id>
-                        <goals>
-                            <goal>generate</goal>
-                        </goals>
-                        <configuration>
-                            <episodes>
-                                <episode>
-                                    <groupId>hu.icellmobilsoft.coffee</groupId>
-                                    <artifactId>coffee-dto-gen</artifactId>
-                                </episode>
-                            </episodes>
-                            <strict>false</strict>
-                            <!-- https://github.com/highsource/maven-jaxb2-plugin/wiki/Catalogs-in-Strict-Mode -->
-                            <catalog>src/main/resources/xsd/hu/icellmobilsoft/quarkus/sampler/dto/super.catalog.xml</catalog>
-                            <schemaIncludes>
-                                <include>xsd/hu/icellmobilsoft/quarkus/sampler/dto/super.xsd</include>
-                            </schemaIncludes>
-                            <bindingIncludes>
-                                <include>xsd/hu/icellmobilsoft/quarkus/sampler/dto/bindings.xjb</include>
-                            </bindingIncludes>
-                            <generateDirectory>${project.build.directory}/generated-sources/src/main/java</generateDirectory>
-                        </configuration>
-                    </execution>
-                </executions>
-                <configuration>
-                    <verbose>true</verbose>
-                    <schemaDirectory>src/main/resources</schemaDirectory>
-                    <args>
-                        <arguments>-openapify</arguments>
-                        <arguments>-Xfluent-api</arguments>
-                        <arguments>-Xannotate</arguments>
-                    </args>
-                    <plugins>
-                        <plugin>
-                            <groupId>hu.icellmobilsoft.jaxb</groupId>
-                            <artifactId>openapi-jaxb-plugin</artifactId>
-                            <version>2.0.0</version>
-                        </plugin>
-                        <plugin>
-                            <groupId>net.java.dev.jaxb2-commons</groupId>
-                            <artifactId>jaxb-fluent-api</artifactId>
-                            <version>2.1.8</version>
-                        </plugin>
-                        <plugin>
-                            <groupId>org.jvnet.jaxb2_commons</groupId>
-                            <artifactId>jaxb2-basics-annotate</artifactId>
-                            <version>1.1.0</version>
-                        </plugin>
-                        <plugin>
-                            <groupId>com.fasterxml.jackson.core</groupId>
-                            <artifactId>jackson-databind</artifactId>
-                            <version>2.14.1</version>
-                        </plugin>
-                    </plugins>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>io.smallrye</groupId>
-                <artifactId>jandex-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>make-index</id>
-                        <goals>
-                            <goal>jandex</goal>
-                        </goals>
-                        <configuration>
-                            <fileSets>
-                                <fileSet>
-                                    <directory>${project.build.directory}/generated-sources/src/main/java</directory>
-                                </fileSet>
-                            </fileSets>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
