--- qs-common/qs-common-dto/src/main/resources/xsd/hu/icellmobilsoft/quarkus/sampler/dto/super.xsd
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<xsd:schema
-            xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-            xmlns="http://super.dto.sampler.quarkus.icellmobilsoft.hu/super"
-            targetNamespace="http://super.dto.sampler.quarkus.icellmobilsoft.hu/super"
-            elementFormDefault="qualified" attributeFormDefault="unqualified"
-            >
-
-    <xsd:import namespace="http://test.dto.sampler.quarkus.icellmobilsoft.hu/test"/>
-    <xsd:import namespace="http://test.dto.sampler.quarkus.icellmobilsoft.hu/post"/>
-</xsd:schema>
