+++ qs-common/qs-common-rest/src/main/resources/i18n/messages.properties
@@ -1,18 +1,18 @@
 
 # classname + key = localized message
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_BAD_REQUEST = The content in the request is not well-formed!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_UNAUTHORIZED = Incorrect credentials!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_FORBIDDEN = The user is not authorized to execute the operation!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_NOT_FOUND = Non-existing endpoint!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_METHOD_NOT_ALLOWED = The endpoint does not support the sent HTTP method!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_NOT_ACCEPTABLE = The endpoint does not support the expected (accept) MIME type!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_UNSUPPORTED_MEDIA_TYPE = The endpoint does not support the sent (content-type) MIME type!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_INTERNAL_SERVER_ERROR = An error occurred during the operation, please try again later!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.REST_SERVICE_UNAVAILABLE = The service currently does not accept requests!
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.INVALID_TYPE_ERROR=Request could not be converted to target object.
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.JSON_PARSE_ERROR=The request sent could not be parsed as a valid JSON.
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.JSON_INVALID_FORMAT_ERROR=Could not deserialize JSON.
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.JSON_MAPPING_ERROR=An error occurred during JSON mapping.
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.HTTP_MESSAGE_ERROR=HTTP message cannot be processed.
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.IO_ERROR=General I/O error.
- hu.icellmobilsoft.quarkus.sampler.dto.exception.enums.FaultType.VALIDATION_ERROR=Validation error.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_BAD_REQUEST=The content in the request is not well-formed!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_UNAUTHORIZED=Incorrect credentials!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_FORBIDDEN=The user is not authorized to execute the operation!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_NOT_FOUND=Non-existing endpoint!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_METHOD_NOT_ALLOWED=The endpoint does not support the sent HTTP method!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_NOT_ACCEPTABLE=The endpoint does not support the expected (accept) MIME type!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_UNSUPPORTED_MEDIA_TYPE=The endpoint does not support the sent (content-type) MIME type!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_INTERNAL_SERVER_ERROR=An error occurred during the operation, please try again later!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.REST_SERVICE_UNAVAILABLE=The service currently does not accept requests!
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.INVALID_TYPE_ERROR=Request could not be converted to target object.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.JSON_PARSE_ERROR=The request sent could not be parsed as a valid JSON.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.JSON_INVALID_FORMAT_ERROR=Could not deserialize JSON.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.JSON_MAPPING_ERROR=An error occurred during JSON mapping.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.HTTP_MESSAGE_ERROR=HTTP message cannot be processed.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.IO_ERROR=General I/O error.
+hu.icellmobilsoft.quarkus.sampler.enums.exception.hu.icellmobilsoft.quarkus.sampler.dto.FaultType.VALIDATION_ERROR=Validation error.
