+++ qs-services/qs-rest-service/src/main/java/hu/icellmobilsoft/quarkus/sampler/rest/restclient/ITestQuarkusSamplerServiceRestRegisteredClient.java
@@ -23,8 +23,8 @@ import jakarta.ws.rs.Path;
 
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
-import hu.icellmobilsoft.quarkus.sampler.api.test.ITestQuarkusSamplerServiceRest;
-import hu.icellmobilsoft.quarkus.sampler.dto.path.QuarkusSamplerPath;
+import hu.icellmobilsoft.quarkus.sampler.api.jakarta.path.QuarkusSamplerPath;
+import hu.icellmobilsoft.quarkus.sampler.api.jakarta.test.ITestQuarkusSamplerServiceRest;
 
 /**
  * REST Registered rest client
